desafio2nf/requirements.txt
streamlit
crewai
crewai-tools
langchain-openai
pandas
python-dotenv
litellm